requirements.txt
Django==1.6.7
djangorestframework==3.1.1
